preparing on started production use

+++ after.sh
@@ -1,5 +1,4 @@
-#!/bin/bash
-
+#!/bin/sh
 # If you would like to do some extra provisioning you may
 # add any commands you wish to this file and they will
 # be run after the Homestead machine is provisioned.
@@ -9,58 +8,90 @@
 # which will be run after this script.
 
 # Install phpmyadmin
-if [[ ! -d "./.etc/phpmyadmin/" ]]; then
+if  [ ! -d ".etc/phpmyadmin" ] && [ ! -f ".etc/phpmyadmin/index.php" ]; then
   sudo apt-get install unzip
   wget https://files.phpmyadmin.net/phpMyAdmin/5.2.1/phpMyAdmin-5.2.1-all-languages.zip
-  unzip phpMyAdmin-5.2.1-all-languages.zip -d ~/noix/.etc/
-  sudo mv ~/noix/.etc/phpMyAdmin-5.2.1-all-languages ~/noix/.etc/phpmyadmin
-  sudo rm -rf phpMyAdmin-5.2.1-all-languages.zip ~/noix/.etc/phpMyAdmin-5.2.1-all-languages
+  unzip phpMyAdmin-5.2.1-all-languages.zip -d ~//noix/.etc/
+  sudo mv ~//noix/.etc/phpMyAdmin-5.2.1-all-languages ~//noix/.etc/phpmyadmin
+  sudo rm -rf phpMyAdmin-5.2.1-all-languages.zip ~//noix/.etc/phpMyAdmin-5.2.1-all-languages
+  touch /etc/phpmyadmin/.gitkeep
 fi
 
+# COPY MYSQL CONFIG
+sudo cp -r /home/vagrant/noix/.etc/mysql/my.cnf /etc/mysql/
+
 # Copy SSL Certificates
-sudo cp -r /etc/ssl/certs/ca.homestead.noix.crt /home/vagrant/noix/.etc/ssl
-sudo cp -r /etc/ssl/certs/ca.homestead.noix.key /home/vagrant/noix/.etc/ssl
-
-# Copy NGINX config
-sudo cp -r /home/vagrant/noix/.etc/nginx/noix.loc /etc/nginx/sites-available/
-
-# Copy Supervisor configs
-#sudo cp -r /home/vagrant/noix/.etc/supervisor/memmon.conf /etc/supervisor/conf.d/
-sudo cp -r /home/vagrant/noix/.etc/supervisor/queue-base.conf /etc/supervisor/conf.d/
-sudo cp -r /home/vagrant/noix/.etc/supervisor/swoole-http.conf /etc/supervisor/conf.d/
+sudo cp -r /etc/ssl/certs/ca.homestead.-api.crt /home/vagrant/noix/.etc/ssl
 
 # Set PHP version
-update-alternatives: using /usr/bin/php8.3
-update-alternatives: using /usr/bin/php-config8.3
-update-alternatives: using /usr/bin/phpize8.3
+sudo update-alternatives --set php /usr/bin/php8.3
+sudo update-alternatives --set phar /usr/bin/phar8.3
+sudo update-alternatives --set phar.phar /usr/bin/phar.phar8.3
 sudo phpenmod xdebug
 
-# Add PPA repository
-sudo add-apt-repository ppa:ondrej/nginx -y
-sudo add-apt-repository ppa:redislabs/redis -y
-
-sudo apt update
-sudo apt upgrade -y
-
 sudo apt install cron -y
-sudo apt install nginx-extras -y
-
-sudo apt install libc-ares-dev libcurl4-openssl-dev -y
-
-pip install --upgrade supervisor
-pip install superlance
-
 sudo apt install php-gmp
 sudo apt install php-bcmath
 sudo apt install php-igbinary
+sudo apt install libc-ares-dev libcurl4-openssl-dev libbrotli-dev
+
+sudo apt remove --purge php5.*-* php7.*-* php8.0-* php8.1-* php8.2-* -y
+sudo rm -rf /etc/php/5.* /etc/php/7.* /etc/php/8.0 /etc/php/8.1 /etc/php/8.2
+
+sudo pecl install yaml
+sudo apt install php8.3-yaml php8.3-amqp php8.3-igbinary php8.3-redis
+
+# Redis
+sudo add-apt-repository ppa:redislabs/redis -y
+
+# CLAMAV
+#sudo apt-get install -y clamav-daemon
+#sudo freshclam
+#sudo systemctl enable --now clamav-daemon clamav-freshclam
+#sudo systemctl enable clamav-daemon
 
 # Install new version beanstalkd, for queue on prod test
 wget https://launchpad.net/ubuntu/+archive/primary/+files/beanstalkd_1.12-2_amd64.deb
 sudo dpkg -i beanstalkd_1.12-2_amd64.deb
 sudo rm -rf beanstalkd_1.12-2_amd64.deb
 
-#cd ~/noix
-#composer config github-oauth.github.com awk -F= '$1 == "APP_URL" {print $2}' .env
-#
+# NGINX
+sudo apt install nginx-extras -y
+sudo add-apt-repository ppa:ondrej/nginx -y
+
+if [ -f ".etc/nginx/-api.loc" ]; then
+  sudo cp -r /home/vagrant/noix/.etc/nginx/-api.loc /etc/nginx/sites-available/
+  sudo cp -r /home/vagrant/noix/.etc/nginx/-api.phpmyadmin.loc /etc/nginx/sites-available/
+fi
+
+# SUPERVISOR
+pip install --upgrade supervisor
+pip install superlance
+
+sudo cp -r /home/vagrant/noix/.etc/supervisor/queue-base.conf /etc/supervisor/conf.d/
+sudo cp -r /home/vagrant/noix/.etc/supervisor/schedule.conf /etc/supervisor/conf.d/
+sudo cp -r /home/vagrant/noix/.etc/supervisor/horizon.conf /etc/supervisor/conf.d/
+
+sudo supervisorctl reread
+sudo supervisorctl restart all
+
+sudo apt update -y
+sudo DEBIAN_FRONTEND=noninteractive apt upgrade -y
+
+sudo apt autoremove -y
+sudo apt autoclean -y
+
+# Project settings commands
+cd /home/vagrant/noix || exit
+
 composer install
-yarn install
+
+php artisan migrate
+php artisan optimize:clear
+php artisan vendor:publish --tag=log-viewer-assets --force
+php artisan l5-swagger:generate
+php artisan cache:clear
+
+cd public || exit
+unlink storage
+ln -s ../storage/app/public storage
